Clippy: Make clippy_project_root work from rust checkout
Before, `clippy_project_root` worked by ascending the path from the build
directory. This worked as long as it was executed from the Clippy
checkout as the Clippy build directory is part of the Clippy checkout.

It didn't work when executed from the rust checkout because the Rust
build directory is at `./rust/build/` which doesn't have the Clippy project
at in the ancestors.

The solution to make the function behave the same in the Rust checkout,
is to ascend from the `CARGO_MANIFEST_DIR` instead.

This commit passed Clippy CI: https://github.com/rust-lang/rust-clippy/pull/6512